+++ autotransform/input/__init__.py
@@ -5,7 +5,7 @@
 # SPDX-License-Identifier: MIT
 # Copyright (c) 2022-present Nathan Rockenbach <http://github.com/nathro>
 
-"""Inputs produce a list of string keys that are used to make changes to the codebase.
-These strings can represent anything and only need to be understood by the components of the Schema.
-Most use cases will return file paths.
+"""Inputs produce a list of Items that are used to make changes to the codebase.
+These Items can represent anything and only need to be understood by the components of the Schema.
+Most use cases will return FileItems.
 """