+++ tests/filter/filter_test.py
@@ -7,17 +7,18 @@
 
 """Utility methods for testing Filter components."""
 
-from typing import Dict
+from typing import List, Tuple
 
 from autotransform.filter.base import Filter
+from autotransform.item.base import Item
 
 
-def run_filter_tests(filt: Filter, test_cases: Dict[str, bool]) -> None:
+def run_filter_tests(filt: Filter, test_cases: List[Tuple[Item, bool]]) -> None:
     """A simple utility method for handling running tests of Filter components
 
     Args:
         filt (Filter): The Filter being tests.
-        test_cases (Dict[str, bool]): A mapping from test case key to expected result.
+        test_cases (List[Tuple[Item, bool]]): A list of test Items with their expected result.
     """
-    for key, result in test_cases.items():
-        assert filt.is_valid(key) == result
+    for item, result in test_cases:
+        assert filt.is_valid(item) == result