+++ tests/filter/test_regex.py
@@ -8,6 +8,7 @@
 """Tests for the RegexFilter component."""
 
 from autotransform.filter.regex import FileContentRegexFilter, RegexFilter
+from autotransform.item.file import FileItem
 
 from .filter_test import run_filter_tests
 
@@ -24,6 +25,7 @@ def test_regex():
         "baz": False,
         "oof": False,
     }
+    test_cases = [(FileItem(path), result) for path, result in test_cases.items()]
     run_filter_tests(filt, test_cases)
 
 def test_inverted_regex():
@@ -38,6 +40,7 @@ def test_inverted_regex():
         "baz": True,
         "oof": True,
     }
+    test_cases = [(FileItem(path), result) for path, result in test_cases.items()]
     run_filter_tests(filt, test_cases)
 
 def test_file_content_regex(tmpdir):
@@ -53,6 +56,7 @@ def test_file_content_regex(tmpdir):
         str(test_file_1): True,
         str(test_file_2): False,
     }
+    test_cases = [(FileItem(path), result) for path, result in test_cases.items()]
     run_filter_tests(filt, test_cases)
 
 def test_inverted_file_content_regex(tmpdir):
@@ -68,4 +72,5 @@ def test_inverted_file_content_regex(tmpdir):
         str(test_file_1): False,
         str(test_file_2): True,
     }
+    test_cases = [(FileItem(path), result) for path, result in test_cases.items()]
     run_filter_tests(filt, test_cases)