+++ tests/input/test_gitgrep.py
@@ -14,15 +14,15 @@ def test_pattern_present() -> None:
     """Tests running git grep when a known match is present."""
 
     inp = GitGrepInput({"pattern": "Tests for the GitGrepInput component."})
-    found_files = inp.get_keys()
+    found_files = inp.get_items()
     assert len(found_files) is 1, "Only one file match expected."
     test_file = __file__.lower().replace("\\", "/")
-    assert found_files[0].lower() == test_file, "The test file should be found."
+    assert found_files[0].get_path().lower() == test_file, "The test file should be found."
 
 
 def test_pattern_not_present() -> None:
     """Tests running git grep when no match is present"""
 
     inp = GitGrepInput({"pattern": "foobar" + "fizzbuzz" + "barfoo" + "buzzfizz" + "notpresent"})
-    found_files = inp.get_keys()
-    assert len(found_files) is 0, "No matches should be found"
+    found_files = inp.get_items()
+    assert len(found_files) == 0, "No matches should be found"